add export excel in activity controller

+++ database/migrations/2024_02_23_052741_create_files_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('tag',100)->nullable();
             $table->string('url')->nullable();
             $table->double('size');
-            $table->tinyInteger('first_version');
+            $table->integer('first_version');
             $table->timestamps();
         });
     }
